requirements.txt
influxdb-client
paho-mqtt
python-telegram-bot
requests
